requirements.txt
mega.py==1.0.8
python-telegram-bot==13.15
loguru==0.6.0
validators~=0.20.0
yt-dlp>=2022.6.29